refactor: update modules with shop admin logic
- update the `EditOrderFeature.js` file by using the current
user's id to determine whether to execute the buyer operation or
the seller operation if the current user is not the super admin.

- update the `RegisterUserFeature.js` file by setting the role of
all new users to `Customer`.

- always emit only the `new:customer` event whenever a new user is
registered.

- implement a `getIsShopAdmin` getter in the `User` model to determine
if a user currently has a record in the `stores` table.

- update the `getIsErrantSeller` getter by returning false if the user
has no record in the `stores` table or the `last_login_time` field is
falsy.